# GH-40720: [Python] Simplify and improve perf of creation of the column names in Table.to_pandas