@jorisvandenbossche 2024-03-21
@github-actions crossbow submit -g python