@pitrou 2024-03-26
> ```python
> In [1]: table = pa.table({'a': [1, 2, 3], 'b': [0.1, 0.2, 0.3], 'c': [3, 4, 5]})
> 
> In [2]: %timeit table.to_pandas()
> 251 µs ± 1.26 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)   # <-- main
> 68.1 µs ± 894 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # <-- PR
> ```

btw, 68 µs still seems non-trivial for such a small table, do we know what the cost factors are?