@jorisvandenbossche 2024-03-26
Based on a cprofile run and for this small example, around 35% comes from the actual conversion in our C++ layer (creating the (copied) numpy arrays), and the remainder is some overhead in our python code, and then mostly the overhead on the pandas' side constructing the objects. 
For example, a large part (ca 20-25%) is in the creation of the Index object for the column names. Looking at the profile, there looks to be some easy room for improvement here, but so that's something to do on the pandas side.